moving config out of exampleDataCreator per request
