+++ ClientSideKeyVaultKeyToCustomerProvidedKey/config.py
