+++ ClientSideKeyVaultKeyToMicrosoftManagedKey/config.py
