+++ ClientSideLocalKeyToCustomerManagedKey/migration.py
@@ -1,4 +1,4 @@
-from ClientSideLocalKeyToCustomerManagedKey.exampleDataCreator import config as cfg
+from ClientSideLocalKeyToCustomerManagedKey import config as cfg
 from azure.storage.blob import BlobServiceClient
 from azure.identity import ClientSecretCredential
 from azure.keyvault.keys import KeyClient
