+++ ClientSideLocalKeyToCustomerProvidedKey/exampleDataCreator/exampleDataCreator.py
@@ -1,6 +1,6 @@
 from azure.storage.blob import BlobServiceClient
 from azure.keyvault.keys import KeyClient
-from ClientSideLocalKeyToCustomerProvidedKey.exampleDataCreator import config as cfg
+from ClientSideLocalKeyToCustomerProvidedKey import config as cfg
 from azure.identity import ClientSecretCredential
