+++ ClientSideLocalKeyToMicrosoftManagedKey/config.py
